test(util): fix uuid spec
