+++ packages/test.ts
@@ -15,7 +15,7 @@ declare const require: any;
 getTestBed().initTestEnvironment(BrowserDynamicTestingModule, platformBrowserDynamicTesting());
 
 // Then we find all the tests.
-// const context = require.context('./', true, /smart-text.spec.ts/);
+// const context = require.context('./', true, /uuid.spec.ts/);
 const context = require.context('./', true, /\.spec\.ts$/);
 // const context = require.context('./', true, /\/chart\/(.*)\.spec\.ts$/);
 // And load the modules.
